A 1.20.1 (neo)forge port (finally works)
