+++ common/src/main/resources/architectury.common.json
@@ -1,0 +1,3 @@
+{
+  "accessWidener": "item_interactions_mod.accesswidener"
+}
