+++ common/src/main/resources/item_interactions_mod.accesswidener
@@ -1,0 +1,1 @@
+accessWidener   v2  named
