+++ forge/src/main/java/dev/mineland/item_interactions_mod/forge/ReloadListenerHelperImpl.java
@@ -1,12 +1,11 @@
-package dev.mineland.item_interactions_mod.neoforge;
+package dev.mineland.item_interactions_mod.forge;
 
 import dev.mineland.item_interactions_mod.Item_interactions_mod;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManagerReloadListener;
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.Mod;
-import net.neoforged.neoforge.client.event.RegisterClientReloadListenersEvent;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
 
 @Mod.EventBusSubscriber(modid = Item_interactions_mod.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
 public class ReloadListenerHelperImpl {
