+++ forge/src/main/resources/META-INF/mods.toml
@@ -1,0 +1,9 @@
+modLoader="javafml"
+loaderVersion = "[47,)"
+license="LGPL-3.0"
+
+[[mods]]
+modId="item_interactions_mod"
+version="2.1.4-(neo)forge"
+displayName="Item Interactions Mod"
+description='''Item interactions in your inventory!'''
