+++ forge/src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+  "pack": {
+    "description": "Example particle for the item interactions mod \n-mineland",
+    "pack_format": 15
+  }
+}
