+++ gradle.properties
@@ -5,9 +5,10 @@ org.gradle.parallel=true
 mod_version=2.1.4
 maven_group=dev.mineland
 archives_name=item-interactions-mod
-enabled_platforms=fabric
+enabled_platforms=fabric,forge
 # Minecraft properties
 minecraft_version=1.20.1
 # Dependencies
 fabric_loader_version=0.16.14
-#neoforge_version=20.2.93
+forge_version=1.20.1-47.4.5
+architectury_version=9.2.14
