--- neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -1,33 +1,0 @@
-modLoader = "javafml"
-loaderVersion = "[2,)"
-#issueTrackerURL = ""
-license = "GNU LGPL 3.0"
-
-[[mods]]
-modId = "item_interactions_mod"
-version = "${version}+${minecraft_version}"
-displayName = "Item interactions mod"
-authors = "mineland"
-description = '''
-Item interactions in your inventory!
-'''
-logoFile = "assets/item_interactions_mod/icon.png"
-
-
-[[dependencies.item_interactions_mod]]
-modId = "neoforge"
-type = "required"
-versionRange = "[21,)"
-ordering = "NONE"
-side = "CLIENT"
-
-[[dependencies.item_interactions_mod]]
-modId = "minecraft"
-type = "required"
-versionRange = "[${minecraft_version},)"
-ordering = "NONE"
-side = "CLIENT"
-
-
-[[mixins]]
-config = "item_interactions_mod.mixins.json"
