+++ settings.gradle
@@ -3,12 +3,16 @@ pluginManagement {
         maven { url "https://maven.fabricmc.net/" }
         maven { url "https://maven.architectury.dev/" }
         maven { url "https://files.minecraftforge.net/maven/" }
+        maven { url = "https://maven.neoforged.net/releases"}
+        maven { url = "https://maven.minecraftforge.net/" }
+        maven { url "https://maven.minecraftforge.net/" }
         gradlePluginPortal()
     }
 }
+
 
 rootProject.name = 'item_interactions_mod'
 
 include 'common'
 include 'fabric'
-//include 'neoforge'
+include 'forge'
